third_party/blink/web_tests/fast/css/invalidation/fieldset-disabled.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<style>
button:disabled { background-color: green }
fieldset + div { color: pink }
</style>
<fieldset id="fieldset">
    <legend></legend>
    <label for="field"></label>
    <input type="text" name="field" id="field">
    <button id=button></button>
    <div>
        <div></div>
        <div></div>
        <div></div>
        <div></div>
    </div>
</fieldset>
<div>
    <div></div>
    <div></div>
    <div></div>
    <div></div>
</div>
<script>
test(() => {
  const transparent = "rgba(0, 0, 0, 0)";
  const green = "rgb(0, 128, 0)";

  assert_equals(getComputedStyle(fieldset, '').backgroundColor, transparent);

  fieldset.offsetTop; // Force recalc.
  fieldset.disabled = true;

  if (window.internals) {
      // There are still instances of SubtreeStyleChange left when updating
      // disabled state. This count should become lower.
      assert_equals(internals.updateStyleAndReturnAffectedElementCount(), 3);
  }

  // Fieldsets can't be disabled or match :disabled, so we have to test a
  // descendant element which becomes disabled by making the fieldset disabled.
  assert_equals(getComputedStyle(button, '').backgroundColor, green);
}, "Use descendant invalidation set for :disabled fieldset.");
</script>
